+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.5.0-dev.2
+- Development release.
+
 ## 0.5.0-dev.1
 - Development release.